gradleの再生成とソースファイルの移動
