+++ settings.gradle
@@ -1,1 +1,2 @@
-rootProject.name = "RayStarkUtils"
+rootProject.name = 'RayStarkUtils'
+
