+++ src/main/java/ray/util/tick/ITickControlled.java
@@ -1,0 +1,5 @@
+package ray.util.tick;
+
+public interface ITickControlled {
+    void done(); //called every tick.
+}
