--- src/ray/util/tick/ITickControlled.java
@@ -1,5 +1,0 @@
-package ray.util.tick;
-
-public interface ITickControlled {
-    void done(); //called every tick.
-}
